# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Jaga Asset Management API",
        "contact": {}
    },
    "paths": {
...
        "/v1/assets": {
            "get": {
                "description": "Retrieve a list of assets with pagination and optional filters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Assets"
                ],
                "summary": "Get list of assets",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GetAssetsResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new asset with the provided details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Assets"
                ],
                "summary": "Create a new asset",
                "parameters": [
                    {
                        "description": "Create asset request",
                        "name": "assetRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateAssetRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.CreateAssetResponse"
                        }
                    }
                }
            }
        },
        "/v1/assets/{id}": {
            "get": {
                "description": "Retrieve an asset by its unique ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Assets"
                ],
                "summary": "Get asset by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Asset ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GetAssetByIDResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update the details of an existing asset by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Assets"
                ],
                "summary": "Update an existing asset",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Asset ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update asset request",
                        "name": "assetRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateAssetRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateAssetResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete an asset by its unique ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Assets"
                ],
                "summary": "Delete an asset",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Asset ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.DeleteAssetResponse"
                        }
                    }
                }
            }
        },
        "/v1/assets/{id}/status": {
            "put": {
                "description": "Update the status of an asset by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Assets"
                ],
                "summary": "Update the status of an asset",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Asset ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update asset status request",
                        "name": "statusRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateAssetStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateAssetResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.AssetDTO": {
            "type": "object",
            "properties": {
                "added_by": {
                    "type": "string"
                },
                "category_id": {
                    "type": "string"
                },
                "category_name": {
                    "type": "string"
                },
                "condition": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "last_maintenance_date": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "purchase_date": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "dto.CreateAssetRequest": {
            "type": "object",
            "required": [
                "added_by",
                "category_id",
                "name",
                "status"
            ],
            "properties": {
                "added_by": {
                    "type": "string"
                },
                "category_id": {
                    "type": "string"
                },
                "condition": {
                    "type": "string",
                    "maxLength": 50
                },
                "last_maintenance_date": {
                    "type": "string"
                },
                "location": {
                    "type": "string",
                    "maxLength": 100
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2
                },
                "purchase_date": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "ready",
                        "under_maintenance",
                        "need_maintenance"
                    ]
                }
            }
        },
        "dto.CreateAssetResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "dto.DeleteAssetResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "dto.GetAssetByIDResponse": {
            "type": "object",
            "properties": {
                "asset": {
                    "$ref": "#/definitions/dto.AssetDTO"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "dto.GetAssetsResponse": {
            "type": "object",
            "properties": {
                "assets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.AssetDTO"
                    }
                },
                "items_per_page": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "page": {
                    "type": "integer"
                },
                "total_items": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
        "dto.UpdateAssetRequest": {
            "type": "object",
            "properties": {
                "added_by": {
                    "type": "string"
                },
                "category_id": {
                    "type": "string"
                },
                "condition": {
                    "type": "string",
                    "maxLength": 50
                },
                "last_maintenance_date": {
                    "type": "string"
                },
                "location": {
                    "type": "string",
                    "maxLength": 100
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2
                },
                "purchase_date": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "ready",
                        "under_maintenance",
                        "need_maintenance"
                    ]
                }
            }
        },
        "dto.UpdateAssetResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "dto.UpdateAssetStatusRequest": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "status": {
                    "type": "string",
                    "enum": [
                        "ready",
                        "under_maintenance",
                        "need_maintenance"
                    ]
                }
            }
        },
...
    }
}